Add theme.select.clear.hover
